requirements.txt
langchain>=0.3.0
langchain-ollama>=0.2.0
langgraph>=0.2.0
pydantic>=2.0
pyyaml>=6.0
httpx>=0.27.0
tavily-python>=0.5.0
ddgs>=7.0.0
